change the whole project structure
