+++ src/Util/CMakeLists.txt
@@ -1,0 +1,3 @@
+cmake_minimum_required(VERSION 3.16.3)
+
+add_library(Util util.c)
