--- src/brainfuckInterpreter.h
@@ -1,31 +1,0 @@
-#pragma once
-#include <string.h>
-
-#define TRUE 1
-#define FALSE 0
-
-#define increment(bfInstance) bfInstance->value++
-#define decrement(bfInstance) bfInstance->value--
-#define DEBUG_COLOR(COL) printf("\x1b[%dm",COL);
-
-
-struct structNode  {
-    int value;
-    int index;
-    struct structNode *previous;
-    struct structNode *next;
-} typedef Node;
-
-size_t getFileLength(const char *path);
-void readFile(const char *path, char *content, const size_t length);
-void interpret(char *brainfuck, char debug);
-char *interpretSection(char *brainfuck, Node **bfInstance, const char debug);
-char decrementPointer(Node **bfInstance);
-void incrementPointer(Node **bfInstance);
-void printDebugMessage(Node *bfInstance);
-char* getNonVisualChars(Node * bfInstance);
-Node *createNewNode();
-void handleDebug(char current,Node* bfInstance, const char debug);
-char isRelevantSymbol(const char symbol);
-void handleDebugSymbol(const char symbol, const char debug);
-void handleDebugMessage(Node *bfInstance, const char debug);
